Add withdraw and update config messages to deposit contract.
