+++ Cargo.toml
@@ -17,5 +17,5 @@ debug-assertions = false
 panic = 'abort'
 incremental = false
 # Please do not disable these. Doing so will cause overflow checks in
-# all workspace members to stop working. Overflows should be errors. 
+# all workspace members to stop working. Overflows should be errors.
 overflow-checks = true
