+++ contracts/cw-pre-propose-base/Cargo.toml
@@ -24,6 +24,7 @@ thiserror = { version = "1.0.31" }
 voting = { version = "*", path = "../../packages/voting" }
 proposal-hooks = { version = "*", path = "../../packages/proposal-hooks" }
 cw-core-interface = { version = "*", path = "../../packages/cw-core-interface" }
+cw-denom = { version = "*", path = "../../packages/cw-denom" }
 
 [dev-dependencies]
 cosmwasm-schema = "1.0.0"
