+++ contracts/cw-pre-propose-base/src/error.rs
@@ -1,4 +1,5 @@
 use cosmwasm_std::StdError;
+use cw_denom::DenomError;
 use thiserror::Error;
 
 use voting::deposit::DepositError;
@@ -9,11 +10,23 @@ pub enum PreProposeError {
     Std(#[from] StdError),
 
     #[error(transparent)]
+    Denom(#[from] DenomError),
+
+    #[error(transparent)]
     Deposit(#[from] DepositError),
 
     #[error("message sender is not proposal module")]
     NotModule {},
 
+    #[error("message sender is not dao")]
+    NotDao {},
+
     #[error("you must be a member of this DAO (have voting power) to create a proposal")]
     NotMember {},
+
+    #[error("no denomination for withdrawal. specify a denomination to withdraw")]
+    NoWithdrawalDenom {},
+
+    #[error("nothing to withdraw")]
+    NothingToWithdraw {},
 }
